requirements.txt
PyMySQL==1.1.1
SQLAlchemy==2.0.40
boto3==1.38.8
botocore==1.38.8
PyJWT==2.6.0